requirements.txt
# Web Scraping
requests>=2.25.1

# Transformação de Dados
tabula-py>=2.3.0
pandas>=1.3.0
numpy>=1.20.0
PyPDF2>=2.0.0

# Banco de Dados
psycopg2-binary>=2.9.1
pymysql>=1.0.2
sqlalchemy>=1.4.0

# API
flask>=2.0.0
flask-cors>=3.0.10

# Utilidades
python-dotenv>=0.19.0